chore: rename android pakage name, applicationId
