+++ android/app/src/main/java/com/ahmatda/newarchitecture/MainApplicationReactNativeHost.java
@@ -1,4 +1,4 @@
-package com.team3app.newarchitecture;
+package com.ahmatda.newarchitecture;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -19,9 +19,9 @@ import com.facebook.react.fabric.CoreComponentsRegistry;
 import com.facebook.react.fabric.FabricJSIModuleProvider;
 import com.facebook.react.fabric.ReactNativeConfig;
 import com.facebook.react.uimanager.ViewManagerRegistry;
-import com.team3app.BuildConfig;
-import com.team3app.newarchitecture.components.MainComponentsRegistry;
-import com.team3app.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
+import com.ahmatda.BuildConfig;
+import com.ahmatda.newarchitecture.components.MainComponentsRegistry;
+import com.ahmatda.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
 import java.util.ArrayList;
 import java.util.List;
