+++ android/app/src/main/java/com/ahmatda/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java
@@ -1,4 +1,4 @@
-package com.team3app.newarchitecture.modules;
+package com.ahmatda.newarchitecture.modules;
 
 import com.facebook.jni.HybridData;
 import com.facebook.react.ReactPackage;
